flask-api/requirements.txt
click==8.0.0
Flask==2.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
PySimpleSOAP==1.16.2
python-dotenv==0.17.1
Werkzeug==2.0.0
xmltodict==0.12.0
